requirements.txt
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.20
django-crispy-forms==1.7.2
django-filter==1.1.0
django-guardian==1.5.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
future==0.17.1
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
Markdown==3.1
MarkupSafe==1.1.1
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2019.1
requests==2.21.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
